Mark all threadpool Workers as weakly referenced (#29)
This is the workaround I mentioned in https://github.com/emscripten-core/emscripten/pull/18227. Since we know all the threads in the app are part of the threadpool, they can be just weakly referenced, so that the existence of the Worker alone doesn't prevent Node.js from exiting, and instead it's the blocking that waits for results of specific ops that keeps the event loop alive.

This allows to get rid of non-JS-esque shutdown helper.